@@ -1,22 +1,24 @@
 {
-  "name": "portfolio4",
+  "name": "portfolio",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@emailjs/browser": "^3.4.0",
-    "@material-ui/core": "^4.12.3",
-    "@material-ui/icons": "^4.11.2",
-    "@testing-library/jest-dom": "^5.16.2",
-    "@testing-library/react": "^12.1.3",
-    "@testing-library/user-event": "^13.5.0",
-    "node-sass": "^7.0.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-router-dom": "^6.2.1",
-    "react-scripts": "5.0.0",
-    "react-type-animation": "^3.2.0",
-    "styled-components": "^5.3.3",
-    "web-vitals": "^2.1.4"
+    "@emailjs/browser": "4.3.3",
+    "@emotion/react": "^11.11.4",
+    "@emotion/styled": "^11.11.5",
+    "@mui/icons-material": "^5.15.19",
+    "@mui/material": "^5.15.19",
+    "@testing-library/jest-dom": "6.4.5",
+    "@testing-library/react": "16.0.0",
+    "@testing-library/user-event": "14.5.2",
+    "node-sass": "9.0.0",
+    "react": "18.3.1",
+    "react-dom": "18.3.1",
+    "react-router-dom": "6.23.1",
+    "react-scripts": "5.0.1",
+    "react-type-animation": "3.2.0",
+    "styled-components": "6.1.11",
+    "web-vitals": "4.0.1"
   },
   "scripts": {
     "start": "react-scripts start",
