@@ -1,11 +1,11 @@
-import React from "react";
-import "./topbar.scss";
-import { Code } from "@material-ui/icons";
+import React from 'react';
+import './topbar.scss';
+import CodeIcon from '@mui/icons-material/Code';
 
 export default function TopBar({ open, setOpen }) {
   return (
     // space here ir important!!
-    <div className={"topBar  " + (open && "active")}>
+    <nav className={'topBar  ' + (open && 'active')}>
       <div className="wrapper">
         <div className="left">
           <a href="#intro" className="logo">
@@ -13,19 +13,19 @@ export default function TopBar({ open, setOpen }) {
           </a>
           <div className="itemContainer">
             <a href="#about" className="logo">
-              <Code className="icon" />
+              <CodeIcon className="icon" />
               <span>About</span>
             </a>
           </div>
           <div className="itemContainer">
             <a href="#projects" className="logo">
-              <Code className="icon" />
+              <CodeIcon className="icon" />
               <span>Work</span>
             </a>
           </div>
           <div className="itemContainer">
             <a href="#contact" className="logo">
-              <Code className="icon" />
+              <CodeIcon className="icon" />
               <span>Contact</span>
             </a>
           </div>
@@ -38,6 +38,6 @@ export default function TopBar({ open, setOpen }) {
           </div>
         </div>
       </div>
-    </div>
+    </nav>
   );
 }
